glusterfsd/mgmt: Connect backup volfile-server in round robin
Once we establish a connection with volfile-server, if that
connection looses, then we try to connect to the ony be one
till the last one on the list. Once we reach the last one,
we never go back to the first one. So if the last one is down
we will fail to connect to that node until it comes back

Change-Id: I8fe801fb536879d13a3c84eb2d935932b44dbaf5
Fixes: #4166
Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>